Can move a player around a 2d world
Simplified camera as well so its thread safe for use in ECS, no more mutable vars

+++ Louganis/Code/Game/Game/EntityDef.h
@@ -17,7 +17,7 @@ public:
     static void LoadFromXML();
     static EntityDef const* GetEntityDef(uint8_t id);
     static EntityDef const* GetEntityDef(std::string const& name);
-    static uint8_t GetEntityDefID(std::string const& name);
+    static int GetEntityDefID(std::string const& name);
 
 private:
